presentation-api/controlling-ua/startNewPresentation_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API, start new presentation tests for Controlling User Agent (success - manual test)</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p>Click the button below and select the available presentation display, to start the manual test.</p>
<button id="presentBtn">Start Presentation Test</button>


<script>
    // check event order
    var order = ['RESOLVE_START', 'FIRE_CONNECTIONAVAILABLE', 'FIRE_CONNECTED'];
    var step = 0;
    // presentation connection
    var connection;
    // disable the timeout function for the tests
    setup({explicit_timeout: true});
    // ----------
    // DOM Object
    // ----------
    var presentBtn = document.getElementById("presentBtn");
    // ---------------------------------------------
    // Start New Presentation Test (success) - begin
    // ---------------------------------------------
    presentBtn.onclick = function () {
        presentBtn.disabled = true;
        promise_test(function (t) {
            var request = new PresentationRequest(presentationUrls);
            return new Promise(function(resolve, reject) {
                request.onconnectionavailable = t.step_func(function(evt) {
                    // Check the order #2: "connectionavailable" event fires
                    assert_equals(order[step], 'FIRE_CONNECTIONAVAILABLE', 'A "connectionavailable" event fires after Promise from PresentationRequest.start() resolves.');
                    step++;
                    assert_equals(connection, evt.connection, 'Both Promise from PresentationRequest() and a "connectionavailable" event handler receive the same presentation connection.');
                });
                request.start()
                    .then(t.step_func(function (c) {
                        connection = c;
                        // Check the order #1: Promise from start() resolves
                        assert_equals(order[step], 'RESOLVE_START', 'Promise from PresentationRequest.start() resolves before connectionavailable event is fired.');
                        step++;
                        // Check the initial state of a presentation connection
                        assert_equals(connection.state, 'connecting', 'The initial state of the presentation connection is "connecting".');
                        // Check the order #3: "connected" event fires
                        connection.onconnect = function() {
                            assert_equals(order[step], 'FIRE_CONNECTED', 'A connect event fires after Promise from PresentationRequest.start() resolves and a "connectionavailable" event fires.');
                            assert_equals(connection.state, 'connected', 'The state of the presentation connection is "connected" when a "connect" event fires.');
                            resolve();
                        };
                        t.step_timeout(function() {
                            reject('A "connect" event did not fire when the presentation connection was established.')
                        }, 5000);
                        // Check, if the connection ID is set
                        assert_true(!!connection.id, 'The connection ID is set.');
                        // Check the type of the connection.id
                        assert_true(typeof connection.id === 'string', 'The connection ID is a string.');
                        // Check the instance of the connection
                        assert_true(connection instanceof PresentationConnection, 'The connection is an instance of PresentationConnection.');
                    }), function(e) { reject(e); });
            });
        }, "The presentation was started successfully.");
    }
    // -------------------------------------------
    // Start New Presentation Test (success) - end
    // -------------------------------------------
</script>
